% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eysenck_2way}
\alias{eysenck_2way}
\title{Howell's data for Eysenck's (1974) memory experiment with 2 independent variables}
\format{
\subsection{\code{eysenck_2way}}{

A data frame with 100 rows and 4 columns:
\describe{
\item{id}{Running subject id}
\item{obs}{Words recalled by a subject in an experiment}
\item{lab}{Label of the experimental condition (factor)}
\item{age}{Age group of the subject (factor)}
}
}
}
\usage{
eysenck_2way
}
\description{
A long-form copy of the data that Howell generated for the chapter 13 of his book on Statistical Methods for Psychology.
}
\keyword{datasets}
