% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howell_eysenck_data}
\alias{howell_eysenck_data}
\title{Howell's data for Eysenck's (1974) memory experiment}
\format{
\subsection{\code{howell_eysenck_data}}{

A data frame with 10 rows and 5 columns:
\describe{
\item{Counting}{Words recalled after counting their letters}
\item{Rhyming}{Words recalled after finding rhyming words}
\item{Adjectives}{Words recalled after finding matching adjectives}
\item{Imagery}{Words recalled after imagining related contexts}
\item{Intentional}{Words recalled after intentional memorizing}
}
}
}
\usage{
howell_eysenck_data
}
\description{
An exact copy of the data that Howell generated for the chapter 11 of his book on Statistical Methods for Psychology.
}
\keyword{datasets}
