% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migraine}
\alias{migraine}
\title{Howell's data on migraine headache and relaxation therapy}
\format{
\subsection{\code{migraine}}{

A data frame with 9 rows and 6 columns:
\describe{
\item{Subject}{Running subject id}
\item{Week1}{Duration (hours) of headaches in week 1 (baseline)}
\item{Week2}{Duration (hours) of headaches in week 2 (baseline)}
\item{Week3}{Duration (hours) of headaches in week 3 (training)}
\item{Week4}{Duration (hours) of headaches in week 4 (training)}
\item{Week5}{Duration (hours) of headaches in week 5 (training)}
}
}
}
\usage{
migraine
}
\description{
A wide-form copy of the data that Howell generated for the chapter 14 of his book on Statistical Methods for Psychology.
}
\keyword{datasets}
