% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{murders}
\alias{murders}
\title{US gun murders by state for 2010}
\format{
\subsection{\code{murders}}{

An object of class "data.frame".
\describe{
\item{state}{US state}
\item{abr}{Abbreviation of US state}
\item{region}{Geographical US region}
\item{population}{State population (2010)}
\item{total}{Number of gun murders in state (2010)}
}
}
}
\usage{
murders
}
\description{
Gun murder data from FBI reports. Also contains the population of each state. Data from dslabs package and Wikipedia.
}
\keyword{datasets}
